[C-981] Add suggested follows to feed if emptyp
