+++ packages/common/src/services/remote-config/feature-flags.ts
@@ -62,10 +62,10 @@ export const flagDefaults: FlagDefaults = {
   [FeatureFlags.WRITE_QUORUM_ENABLED]: false,
   [FeatureFlags.EARLY_ACCESS]: false,
   [FeatureFlags.SUPPORTER_DETHRONED_ENABLED]: false,
-  [FeatureFlags.PLAYLIST_ENTITY_MANAGER_ENABLED]: false,
-  [FeatureFlags.SOCIAL_FEATURE_ENTITY_MANAGER_ENABLED]: false,
-  [FeatureFlags.TRACK_ENTITY_MANAGER_ENABLED]: false,
-  [FeatureFlags.USER_ENTITY_MANAGER_ENABLED]: false,
+  [FeatureFlags.PLAYLIST_ENTITY_MANAGER_ENABLED]: true,
+  [FeatureFlags.SOCIAL_FEATURE_ENTITY_MANAGER_ENABLED]: true,
+  [FeatureFlags.TRACK_ENTITY_MANAGER_ENABLED]: true,
+  [FeatureFlags.USER_ENTITY_MANAGER_ENABLED]: true,
   [FeatureFlags.NEW_ARTIST_DASHBOARD_TABLE]: false,
   [FeatureFlags.BUY_AUDIO_ENABLED]: false,
   [FeatureFlags.BUY_AUDIO_COINBASE_ENABLED]: false,
