+++ packages/mobile/src/components/suggested-follows/index.ts
@@ -1,0 +1,1 @@
+export * from './SuggestedFollows'
