+++ packages/mobile/src/screens/feed-screen/FeedScreen.tsx
@@ -6,12 +6,17 @@ import {
   feedPageLineupActions as feedActions,
   feedPageSelectors
 } from '@audius/common'
-import { useDispatch } from 'react-redux'
+import type { FollowArtists } from 'audius-client/src/common/store/pages/signon/types'
+import * as signOnActions from 'common/store/pages/signon/actions'
+import { getFollowArtists } from 'common/store/pages/signon/selectors'
+import { Dimensions, View } from 'react-native'
+import { useDispatch, useSelector } from 'react-redux'
 
 import IconFeed from 'app/assets/images/iconFeed.svg'
 import { Screen, ScreenContent, ScreenHeader } from 'app/components/core'
 import { Lineup } from 'app/components/lineup'
 import { OnlineOnly } from 'app/components/offline-placeholder/OnlineOnly'
+import { SuggestedFollows } from 'app/components/suggested-follows'
 import { usePopToTopOnDrawerOpen } from 'app/hooks/usePopToTopOnDrawerOpen'
 import { make, track } from 'app/services/analytics'
 
@@ -21,8 +26,11 @@ const { makeGetLineupMetadatas } = lineupSelectors
 
 const getFeedLineup = makeGetLineupMetadatas(getDiscoverFeedLineup)
 
+const dimensions = Dimensions.get('window')
+
 const messages = {
-  header: 'Your Feed'
+  header: 'Your Feed',
+  emptyFeed: `Oops! There's nothing here.`
 }
 
 export const FeedScreen = () => {
@@ -37,6 +45,12 @@ export const FeedScreen = () => {
     },
     [dispatch]
   )
+
+  const followArtists: FollowArtists = useSelector(getFollowArtists)
+  const onPressFollow = () => {
+    dispatch(signOnActions.followArtists(followArtists.selectedUserIds))
+    dispatch(feedActions.fetchLineupMetadatas())
+  }
 
   return (
     <Screen>
@@ -55,6 +69,14 @@ export const FeedScreen = () => {
           pullToRefresh
           delineate
           selfLoad
+          LineupEmptyComponent={
+            <View style={{ height: dimensions.height - 200 }}>
+              <SuggestedFollows
+                title={messages.emptyFeed}
+                onPress={onPressFollow}
+              />
+            </View>
+          }
           actions={feedActions}
           lineupSelector={getFeedLineup}
           loadMore={loadMore}
